[#1055] Fix cast operation from TIMESTAMP to TIMESTAMP WITH TIME ZONE
The changes in 22ab0173 failed to consider the possibility of TIMESTAMP being cast to TIMESTAMP WITH TIME ZONE. Following is the output difference
for the query between Octo and Postgres.
```sql
-- Postgres
names=> select timestamp'2023-08-01 00:00:00'::timestamp with time zone;
      timestamptz
------------------------
 2023-08-01 00:00:00-04
(1 row)

-- Octo
OCTO> select timestamp'2023-08-01 00:00:00'::timestamp with time zone;
TIMESTAMPTZ
2023-08-01 01:00:00-04
(1 row)
```
Similar to 22ab0173 this issue is because `tm->tm_isdst` is not set to `-1` before calling `mktime()` in `ydboctoDateTimeCastC()`.
This is fixed in this commit by adding the missing initialization in the external function for
the case when a TIMESTAMP is being casted to TIMESTAMP WITH TIME ZONE.

TDTT099 is updated to validate this fix.