[#382] Fix date/time internal representation when the value being represented is before EPOCH or is of type TIME
Issue
-----------
Queries like the following give wrong result in Octo
```sql
select timestamp'0001-01-01 01:01:01' < timestamp'0001-01-01 01:01:01.222';
select time'0:37:37' between time'0:37:37.412205' and time'6:16:21';
```
They are tracked by https://gitlab.com/YottaDB/DBMS/YDBOcto/-/issues/382#note_2225026843 and
https://gitlab.com/YottaDB/DBMS/YDBOcto/-/issues/382#note_2220821669 and need to be fixed.

Analysis
--------
Consider a timestamp query similar to the one mentioned in **issue** section:
```sql
OCTO> select timestamp'1969-12-31 23:59:59.999998'<timestamp'1969-12-31 23:59:59.999999';
f
(1 row)
```
In Octo, `-1999998` represents `timestamp'1969-12-31 23:59:59.999998'` and `-1999999` represents
`timestamp'1969-12-31 23:59:59.99999`. M plan compares the internal representations to evaluate the `<` expression.

Here `-1999998` represents a value which is not less than `-1999999` but
`timestamp'1969-12-31 23:59:59.999998'` (represented by `-1999998`) is less than `timestamp'1969-12-31 23:59:59.99999`
(represented by `-1999999`).

The problem here is the way microseconds are represented in date/time values that are before EPOCH.

The internal format of date/time value is `[seconds][microseconds]`. The last 6 digits always represent [microseconds].

Date/time inputs that are less than EPOCH (`1970-01-01 00:00:00`) have negative values in Octo's internal notation for
example `1969-12-31 23:59:59.999999` is `-1999999`.

In `-1999999`, `-1` represents seconds since epoch and `999999` is microseconds.

This form of internal format was chosen during implementation of #382 because mktime() works with only seconds so microseconds had
to be handled by Octo. Octo handles it by maintaining microseconds and seconds separately within a long integer. The last 6 digits
represents microseconds and the rest represent seconds.

The internal format value is formed by copying the microseconds value as the last 6 digits irrespective of whether the seconds part
was negative or not. This leads to wrong comparisons between date/time values as explained above.

The internal representation computation logic should be fixed to take into account whether the seconds part of the internal format
is negative or not. If negative then the microseconds value needs to be converted to be its nine's compliment.

When nine's compliment is applied to the microsecond value `timestamp'1969-12-31 23:59:59.999998'` and
`timestamp'1969-12-31 23:59:59.99999` will internally be represented as `-1000001` and `-1000000` and this will fix the comparison
operations.

Nine's compliment represents a `000000` microsecond value as `999999` and `999999` microsecond value as `000000`. It is obtained
by subtracting microseconds value by `999999`. Note microseconds always is 6 digits so 6 9's are used to get nine's compliment.

A ten's compliment which is `(999999-microseconds)+1` causes `000000` microsecond value to be `1000000`. Here even if the most
significant digit is ignored queries like the following which evaluate to `-1000000 < -1000001` will result in wrong values.
The same expression using a nine's compliment will be `-1999999` < `-1000000` and this leads to the correct result. So
nine's compliment is chosen.
```
select timestamp'1969-12-31 23:59:59'<timestamp'1969-12-31 23:59:59.999999';
```

The same logic applies to `time` data. They also have a similar problem as explained in
https://gitlab.com/YottaDB/DBMS/YDBOcto/-/issues/382#note_2220821669. The reason why `time` has this issue is because mktime()
for time returns a negative value. This results in the issue we discussed previously when we try to add microseconds.

Change
-------
* A `NINES_COMPLIMENT()` macro is added to compute the compliment of a given value
* `ydboctoText2InternalFormatC()` in src/aux/ydboctodateoperations.c is updated to convert microseconds to nine's compliment when
  result of `mktime()` is -ve
* Rest of the date/time functions make use of `ADD_MICRO_SECONDS()` and `REMOVE_MICRO_SECONDS()` to handle seconds and
  microseconds separation and formation to and from internal representation. These have been updated to make use of
  `NINES_COMPLIMENT()`.
* To form the various date/time formats supported by Octo and to perform date/time computations other than comparison we need the
  original microseconds value. To get this we re-apply nine's compliment on the value which is nine's compliment of the original
  value. This is done by `REMOVE_MICRO_SECONDS()`.
* Since internal representation to ZUT conversion logic in `PrintDateTimeResultColumnValue()` of src/aux/_ydboctoplanhelpers.m
  doesn't make use of the above mentioned macros, the nines compliment logic is added here as well to get the actual micro seconds
  from internal representation.
* With this change `xrefs` of readonly tables will have to be recreated as the date/time internal representation values in the
  `xrefs` generated previous to this change will be out of date.
  Bumping `FMT_PLAN_DEFINITION` from `33` to `34` takes care of this as auto upgrade procedure of Octo calls `DISCARD ALL` on seeing
  `FMT_PLAN_DEFINITION` change and this causes `xrefs` to be discarded and regenerated as and when needed. In case of readwrite
  tables we don't need to worry about it as the usecase that is fixed is not known to be used at this time.

Test
------
* tests/fixtures/TDTT099.sql is updated to include the queries which lead to this MR
* The change in internal format representation caused tests/outref/TDT04.ref, tests/outref/TDA01.ref and tests/outref/TC030.ref
  to be updated. The change replaces old values with the new ones that will be seen after this change.