[#382] Fix timestamp processing at the beginning of daylight savings when time zone is not in the literal
Issue
-------
This issue was noticed while working on #845 (Interval). A timestamp at the beginning of daylight savings was not processed
correctly.

The beginning of daylight savings is the second Sunday of March which in the year `2024` is on `2024-03-10` and at time `02:00:00`.
This means on `2024-03-10` the clock is set forward by 1 hr at `02:00:00`. So after `01:59:59` we see `03:00:00`.

In Octo, when a select on such timestamp is executed we get the following
```sql
OCTO> select timestamp with time zone'2024-03-10 02:00:00';
???
2024-03-10 01:00:00-05
(1 row)
```

This is incorrect as `02:00:00` is forwarded to `03:00:00` as shown below
```sql
names=> select timestamp with time zone'2024-03-10 02:00:00';
      timestamptz
------------------------
 2024-03-10 03:00:00-04
(1 row)
```

When timezone is specified in the literal, Octo's processing is correct
```sql
OCTO> select timestamp with time zone'2024-03-10 02:00:00-05';
???
2024-03-10 03:00:00-04
(1 row)
OCTO> select timestamp with time zone'2024-03-10 02:00:00-04';
???
2024-03-10 01:00:00-05
(1 row)
```
So, Octo's processing at the beginning of DST when no timezone is specified needs to be fixed.

Analysis
---------
In the query's M plan, `$$Transform2UnixTime^%ydboctoplanhelpers` is called to convert date/time text literal to internal
representation. To do this `ydboctoText2InternalFormatC()` in src/aux/ydboctodateoperations.c is invoked. This is where we need to
look.

In it, strptime() is run to get `struct tm` values for `timestamp with time zone'2024-03-10 02:00:00'`
```
$3 = {tm_sec = 0, tm_min = 0, tm_hour = 2, tm_mday = 10, tm_mon = 2, tm_year = 124, tm_wday = 0, tm_yday = 0, tm_isdst = 0, __tm_gmtoff = 0, __tm_zone = 0x0}
```

Code block at line 1617 below gets the timezone info as its not present in the literal.

**src/aux/ydboctodateoperations.c**
```
1615                 } else {
1616                         // Factor in local timezone value
1617                         struct tm tm2;
1618                         memcpy(&tm2, &tm1, sizeof(struct tm));
1619                         tm2.tm_isdst = -1;
1620                         mktime(&tm2);
1621                         gmtoff = (-1) * tm2.__tm_gmtoff;
1622                 }
1623                 tm1.tm_sec += gmtoff; // Local time zone to UTC
1624         }
```
This is done by copying `tm1` (original holder of input) to `tm2` (tmp structure) and setting `tm2.tm_isdst` to `-1` such that
mktime() at 1620 can determine daylight savings info.

`tm2` value before and after mktime() is shown below.
```
Before
$9 = {tm_sec = 0, tm_min = 0, tm_hour = 2, tm_mday = 10, tm_mon = 2, tm_year = 124, tm_wday = 0, tm_yday = 0, tm_isdst = -1, __tm_gmtoff = 0, __tm_zone = 0x0}
After
$10 = {tm_sec = 0, tm_min = 0, tm_hour = 3, tm_mday = 10, tm_mon = 2, tm_year = 124, tm_wday = 0, tm_yday = 69, tm_isdst = 1, __tm_gmtoff = -14400, __tm_zone = 0x555555dd4e90 "EDT"}
```
You see `__tm_gmtoff` is `-14400` and `tm_hour` is set to 3. This is because the given time which is `02:00:00` for the particular
day is not present and the time is adjusted as shown above but this structure is used temporarily to just
get the timezone. The timezone retrieved is applied on the original structure `tm1`. Which has the following values.
```
$11 = {tm_sec = 0, tm_min = 0, tm_hour = 2, tm_mday = 10, tm_mon = 2, tm_year = 124, tm_wday = 0, tm_yday = 0, tm_isdst = 0, __tm_gmtoff = 0, __tm_zone = 0x0}
```
This is incorrect and needs to be fixed as other DST related changes done to `tm` structure needs to be taken into account in
addition to timezone value.

Rest of the computation with the original value (`05` is considered as the local time zone at the time of this commit):
```
Computation to reach UTC as we store in UTC
02:00:00 + 04 = 06:00:00 UTF

Computation to reach back from UTC to local time
06:00:00 UTF - 05 = 01:00:00
```
This is how we see `01:00:00` in the output without this change.

Rest of computation With the correction:
```
Computation to reach UTC as we store in UTC
03:00:00 + 04 = 07:00:00 UTF

Computation to reach back from UTC to local time
07:00:00 -05 = 02:00:00 since 02:00:00 doesn't exist for the date this is converted to 03:00:00 by localtime()

Suppose -04 was used (if we are executing this query when DST was active)
07:00:00 -04 = 03:00:00 since 03:00:00 exists for the date this is not changed by localtime()
```
As seen above the correction proposed will fix this issue. Note, mktime() call at line 1620 to obtain timezone info will give
`-14400` i.e. `-04` whether the query is run with localtime being in EST or EDT because of the date present in the input.

Fix
-----
* src/aux/ydboctodateoperations.c code explained in **analysis** section is updated to make use of `tm1` itself for the mktime()
  call. This ensures the DST changes are reflected in the `struct tm` values leading to the correct output. This fixed the issue.

Test
------
* TDTT079 is updated to cross check the query which validate this fix