[#382] Compare TIMESTAMP WITH TIME ZONE in UTC
Issue
------
While analyzing a cron job failure of TDTT087/TDTT088 following query was noticed to have different result from Postgres
```sql
names=> select timestamp with time zone'2024-11-03 01:49:19-04' = timestamp with time zone'2024-11-03 01:49:19-05';
 ?column?
----------
 f
(1 row)

OCTO> select timestamp with time zone'2024-11-03 01:49:19-04' = timestamp with time zone'2024-11-03 01:49:19-05';
???
t
(1 row)
```

This is incorrect as `timestamp with time zone'2024-11-03 01:49:19-04'` is not the same as
`timestamp with time zone'2024-11-03 01:49:19-05'`. They differ by 1 hour. This needs to be fixed in Octo.

Analysis
----------
All Day Light Savings End Time periods have this issue in Octo.

DST change in 2024 is explained below
* Daylight saving time began in 2024 on Sunday, March 10 at 2 a.m. local time, when our clocks moved forward an hour
* Daylight saving time will end for the year 2024 at 2 a.m. local time on Sunday, Nov. 3, when we "fall back" and gain an extra hour of sleep

Query:
```sql
select timestamp with time zone'2024-11-03 01:49:19-04' = timestamp with time zone'2024-11-03 01:49:19-05';
```

Relavant parts of the M plan generated is shown below
```
.
.
 17     . SET %ydboctoexpr=1
 18     . DO:%ydboctoexpr
 19     . . SET:1&$INCREMENT(%ydboctocursor(cursorId,"keys",3,"","")) %ydboctoz=42
 20     . . SET %ydboctoexpr=$$Transform2UnixTime^%ydboctoplanhelpers($GET(%ydboctocursor(cursorId,"parameters",1)),21,36,"%Y-%m-%d %H:%M:%S%z")
 21     . . SET %ydboctoexpr=$$ConvertToLocalTimezone^%ydboctoplanhelpers(21,%ydboctoexpr)
 22     . . SET %ydboctoexpr(1)=$$ForceNumeric^%ydboctoplanhelpers(%ydboctoexpr)
 23     . . SET %ydboctoexpr=$$Transform2UnixTime^%ydboctoplanhelpers($GET(%ydboctocursor(cursorId,"parameters",2)),21,36,"%Y-%m-%d %H:%M:%S%z")
 24     . . SET %ydboctoexpr=$$ForceNumeric^%ydboctoplanhelpers(%ydboctoexpr)
 25     . . SET %ydboctoexpr=$$ConvertToLocalTimezone^%ydboctoplanhelpers(21,%ydboctoexpr)
 26     . . SET %ydboctoexpr=%ydboctoexpr(1)=%ydboctoexpr
 27     . . SET %ydboctoexpr=$select((1=%ydboctoexpr):"t",(0=%ydboctoexpr):"f",1:$ZYSQLNULL)
.
.
```

Lets look into different intermediate data created (microseconds are separated from seconds in IR (internal representation) values below by a space
just to see the value of concern clearly.
Also, ignore the microseconds in IR for this usecase as its not relevant for this usecase)
* 1st operand (`2024-11-03 01:49:19-04`) goes through conversion to IR in $Transform2UnixTime() (line 20). Result IR:`1730612959 000000`
  * The seconds of the IR value is in UTC (timezone -> +00) and is obtained by adding 14400 to seconds since epoch representation of `2024-11-03 01:49:19`
* Then it is converted to localtime by using $ConvertToLocalTimezone in line 21. Result: IR:`1730598559 000000`
  * The result is in local timezone and is obtained by adding -14400 to UTC value
* 2nd operand (`2024-11-03 01:49:19-05`) goes through conversion to IR in $Transform2UnixTime() (line 23). Result IR:`1730616559 000000`
  * The seconds of the IR value is in UTC and is obtained by adding 18000 to seconds since epoch representation of `2024-11-03 01:49:19`
* Then it is converted to localtime by using $ConvertToLocalTimezone in line 25. Result IR:`1730598559 000000`
  * The result is in local time zone and is obtained by adding back -18000 to UTC value
* Finally, they are compared. Result: `t`.

Difference before converting to local time zone is 3600 seconds or 1 hour

Difference after converting to local time zone is 0 seconds or 0 hours

In case of the query in question both operands are `timestamp with time zone` and this means its IR value which is in UTC would have factored in
timezone and daylight saving information. So comparing the UTC values will give a valid result and avoid the problems that occur
with converting to local time zone.

The only time conversion to localtime zone of a `timestamp with time zone` value is needed is when it is being compared with a date/time
value which doesn't have time zone. For such types that don't have time zones the UTC value doesn't factor in any time zone information
i.e. the seconds from EPOCH in local time zone and in UTC is same where as for types with time zone its time zone value is added to obtain
its UTC value. So the type with time zone has to be converted to local time zone when it is compared with a type without time zone.

Why we need to do the comparison using IR values is because this helps to do key fixing optimization by passing the IR value as a subscript
to cross referenced globals. tm structure can also be used to do comparison but key fixing wouldn't be easily possible as it has several fields.

Change
--------
* src/m_templates/tmpl_print_expression.ctemplate is updated to avoid adding `$ConvertToLocalTimezone` when the comparison is between
  two `TIMESTAMP WITH TIME ZONE` operands. For IN operation the IN list is ensured to have `TIMESTAMP WITH TIME ZONE` type nodes
  if the left operand is of that type. So, code related to IN operation is updated to check if first operand is of the concerning type
  and if so skip addition of `$ConvertToLocalTimezone` for all of the nodes in the IN list.
* IS_COMPARE_OP(plan_type) which is used in the above change is also used to remove a switch statement checking the same set of
  conditions in src/m_templates/tmpl_print_expression.ctemplate.

Test
-------
* tests/fixtures/TDTT079.sql is updated to have queries which validate this fix