[#1080] Fix rare TDTT088 failures
Summary
-------

TDTT088 would fail in 1/4000 runs. This failure was found to be due to
converting a timestamp with time zone to Horolog format (which doesn't
have a timezone), and then converting it back to a timestamp with time
zone format. Normally, there is no problem with this conversion, as long
as the timezone doesn't change. But if the random date generated falls
in the middle of the Autumn "fall back" change (in EST, that's between
1am and 2am on the first Sunday in November), the timezone is ambigious,
times between 1am and 2am are traversed twice in the clock. The fix is
to set the test to a Timezone in which DST does not happen. UTC is such
a timezone.

Details
-------
Here's how this bug happens step by step:

1. Create these two tables:
```
create table timestamptz_horolog_tbl (order_id integer primary key, order_timestamptz timestamp(horolog) with time zone) GLOBAL "^timestamptzhorologtbl" READWRITE;
create table timestamptz_zhorolog_tbl (order_id integer primary key, order_timestamptz timestamp(zhorolog) with time zone) GLOBAL "^timestamptzzhorologtbl" READWRITE;
```

2. Run these two queries, and then produce output from them:
```
select timestamptz_to_horolog(timestamp with time zone '2036-11-02t03:41:58-02:29'); -> 71529,4258 <<
select timestamptz_to_zhorolog(timestamp with time zone '2036-11-02t03:41:58-02:29'); -> 71529,4258,,18000
```

3. Insert the output into tables:
```
insert into timestamptz_horolog_tbl  values (19, timestamp(horolog) with time zone '71529,4258'); <<
insert into timestamptz_zhorolog_tbl values (20, timestamp(zhorolog) with time zone '71529,4258,,18000');
```

4. Select from Tables:
```
OCTO> select * from timestamptz_horolog_tbl;
order_id|order_timestamptz
19|2036-11-02 01:10:58-04 <<

OCTO> select * from timestamptz_zhorolog_tbl;
order_id|order_timestamptz
20|2036-11-02 01:10:58-05
```

For the query that this test relies on to succeed, the outputs need to be
identical.

Note how the output for the Horolog conversion lost the timezone, and when we
try to add a timezone back based on the current timezone, we put the wrong
timezone on. This is because if the random date generated falls
in the middle of the Autumn "fall back" change (in EST, that's between
1am and 2am on the first Sunday in November), the timezone is ambigious,
as the times between 1am and 2am are traversed twice in the clock. So the
output for horolog above could as correctly be in the `-04` or `-05`
offset.

Fix
---
In the test, set the timezone to be a timezone which does not have DST
changes. Setting the test to use the UTC timezone solves the problem.