Fix TYIO01 subtest to not modify reference file in the repository
Test TYIO01 modified the reference file `tests/outref/TYIO01.ref` on the
fly. This led to the file being modified in the source tree and then
inadvertently committed in the source during a git commit. See
https://gitlab.com/YottaDB/DBMS/YDBOcto/-/merge_requests/1632#note_2599726672
for an instance for when this happened.

The commit changes the test to modify a copy of the reference file, not
the actual reference file. `verify_output()` in `test_helpers.bash` now
checks to see if a copy of the reference file is present; if so, then
use that instead of the reference file.