[#1086] Issue error for ORDER BY/GROUP BY TRUE/FALSE
During testing with Ubuntu 24.04, which has Postgres 16, Octo tests
`test_order_by/TOB17` and `test_group_by/TGB20` started failing.
Investigating the cause of this failure showed that it was due to
queries cross checked on Postgres that passed with Postgres 14 now
starting to fail with Postgres 16. The queries that failed are:

1. ORDER BY BOOLEAN
2. GROUP BY BOOLEAN

Sample error on Postgres 16:

```
names=> select firstname from names order by TRUE;
ERROR:  non-integer constant in ORDER BY
LINE 1: select firstname from names order by TRUE;
```

In Postgres 14, this was the output:
```
names=> select firstname from names order by TRUE;
 firstname
-----------
 Zero
 Acid
 Cereal
 Lord
 Joey
 Zero
(6 rows)
```

Order by/Group by boolean had no effect and was done in Octo purely for
compatibility with Postgres. The Octo source code has now been updated
to disallow such expressions.

The only change done is in `src/qualify_statement.c`. While at it, I
noticed that `error_encountered` was barely used, so I removed it to
simplify the code.

The tests were refactored so that all such uses went from the tests that
are expected to run successfully to the tests that run the query and
expect errors.

I have been unable to find the exact commit in Postgres that changes its
behavior, as it's not documented in the release notes for 15 or 16.