[YottaDB/DB/YDB#1148] Fix test-auto-upgrade job failure due to syntax error in tools/ci/build.sh (regression in 624a5585)
The previous commit 624a55857c76ec073dcb9a5749361dae996a818a tried to
fix this expression which didn't work on a newer version of bash `5.2.21`:

```
fixtures="$('ls ../tests/*.bats.in ../tests/fixtures/!(TC058*|TC059*|TC060*).sql')"
```

This failed in a pipeline in the auto-upgrade job:
https://gitlab.com/YottaDB/DBMS/YDBOcto/-/jobs/10708080927

The problem are the single quotes inside the `$()` expression. Instead
of using this expression, I rewrote the expression using `find` as it's
easier to maintain in the future.

It was locally tested using this command:

```
docker run -it --link mysql --rm tests-ubuntu "test-auto-upgrade" "debug" "8587b12086666c88ea2c8a19b55a736629269907"
```