[Test] Fix rare failure in TDTT087
This issue is identical to !1638 (4d86e565). See that MR/commit for more
details. The gist of the issue is that a random date got generated that
falls during the "Fallback" DST switch. Such a date is ambiguous, and
cannot be resolved correctly between some types.

In the process of fixing TDTT087 and using an Asia/Bangkok timezone
instead of UTC to add a variety of conditions to the test system
(neither one does Daylight Savings Time (DST)), TDTT087 started failing
reliably with another symptom:

```
Comparing outref/TDTT087.ref output.txt
32c32
< t
---
> f
```

This is because this query returned false (without the count, it returns
0 results):

```
select count(*) >= 10 from timetz_horolog_tbl t1 inner join timetz_zhorolog_tbl t2 on t1.order_timetz = t2.order_timetz;
```

A lot of examination of the underlying data and workflow of the test
finally revealed that the bug is in this function:

```
$ TZ="Asia/Tokyo" octo
OCTO> select timetz_to_zhorolog(time with time zone'10:00:00');
timetz_to_zhorolog
,36000,,32400
(1 row)
OCTO>
```

The timezone for `$ZHOROLOG` is supposed to be actually -32400, not 32400.

A sample correct behavior for another function is as follows:

```
$ TZ="Asia/Tokyo" octo
OCTO> select timestamptz_to_zhorolog(now());
timestamptz_to_zhorolog
67454,10311,463356,-32400
(1 row)
```

Notice that the timezone is correct here at -32400.

Code analysis showed that the bug was actually a copy-pasta bug--it
looks like the code was copied from an earlier mostly identical block
but the variables not changed. In `src/aux/_ydboctoplanhelpers.m`,
`timestampWithTimeZone` was supposed to be `timeWithTimeZone`, as in
that block that's the variable that is being used.

With that change, the test passes now reliably.